+++ web/src/App.tsx
@@ -6,6 +6,8 @@ import { EnergyPilot } from './pages/EnergyPilot'
 import { KonfigurationGlobal } from './pages/KonfigurationGlobal'
 import { KonfigurationGeraete } from './pages/KonfigurationGeraete'
 import { KonfigurationGeraet } from './pages/KonfigurationGeraet'
+import { SensorenUeberschuss } from './pages/SensorenUeberschuss'
+import { SensorenHausbilanz } from './pages/SensorenHausbilanz'
 
 /* Ausschliesslich die Routentabelle. Das Layout ist Elternroute mit <Outlet />,
    damit Navigation und Kopfzeile beim Seitenwechsel nicht neu montiert werden.
@@ -19,6 +21,8 @@ export function App() {
         <Route path="/" element={<Status />} />
         <Route path="/steuerung" element={<Steuerung />} />
         <Route path="/energy-pilot" element={<EnergyPilot />} />
+        <Route path="/sensoren/ueberschuss" element={<SensorenUeberschuss />} />
+        <Route path="/sensoren/hausbilanz" element={<SensorenHausbilanz />} />
         <Route path="/konfiguration/global" element={<KonfigurationGlobal />} />
         <Route path="/konfiguration/geraete" element={<KonfigurationGeraete />} />
         <Route path="/konfiguration/geraete/neu" element={<KonfigurationGeraet mode="create" />} />