+++ web/src/pages/SensorenHausbilanz.tsx
@@ -1,0 +1,5 @@
+import { SensorFormulaForm } from '../components/SensorFormulaForm'
+
+export function SensorenHausbilanz() {
+  return <SensorFormulaForm kind="battery_residual" title="Hausbilanz" />
+}